SystemUI: open settings panel if there are no notifications
If there are no notifications that belong to the current user, then automatically open the quick settings menu (like the two-finger open). It only ignores the ADB icon for now (which is pretty useless) but could be extended to ignore the usb mounting one or maybe another?

TODO: add Settings.System toggle to disable the option

Change-Id: I8c53c3323579f1a2136c6c2309351298e7ebb46f
Signed-off-by: Roman Birg <romanbirg@gmail.com>
Signed-off-by: iceandfire <arhamjamal@gmail.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java