Forward properly setForceUse requests from AudioService to Policy
There is a potential race condition when toggling between speaker and BT_SCO
through setForceUse calls. This is because mForcedUseForComm variable is not
thread-safe and depending on call sequence, we may result in routing to
wrong output device.

This patch introduces a new method, reconsiderUsageForComm(), which will
reconsider the forced config switch in order to send the proper request
according to SCO/Speaker state.

Change-Id: Ibd53e6cf49e5277c2f0bea5fc5f64ae9379f9f08
Signed-off-by: Piotr Diop <piotrx.diop@intel.com>
Signed-off-by: David Wagner <david.wagner@intel.com
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 100472
Signed-off-by: pcarenza <pcarenza@gmail.com>