Fix graphics corruption caused by non-uniqueness of a resource address
Some caches(PatchCache, TextureCache, PathCache) for HWUI uses deferred
removal for their cache entries even though actual resource objects
are already freed by ResourceCache.
For this reason, the uniqueness of a resource address is not guaranteed
in specific cases. (When malloc() and free() are called very frequently,
malloc() can return the same address.)

For example, it can be possible that a cache have two cache entries
for two different resources but the same memory address.
(Of course one of the resources is already freed.)
It also can be possible mGarbage vector in PatchCache has duplicated
addresses and this can lead to duplicated free blocks in the free
block list and graphics corruption.

So this patch makes sure resource objects are freed after the resources
are removed from the caches to guarantee the uniqueness of a resource
address.

Change-Id: I040f033a4fc783d2c4bc04b113589657c36fb15b
Signed-off-by: Sangkyu Lee <sk82.lee@lge.com>