Frameworks: Update Google dialer check for QS 3.0
As the network tile is now on the back of the rssi tile instead of it's own tile,
it's no longer feasible to simply hide the network tile if the google dialer is installed.

Quick setting tiles are created BEFORE boot, but we can only confirm the dilaer type AFTER
boot. Instead of recreating the rssi tile as a single tile, simply disable the
on click network mode toggle if google dialer is present and provide an appropriate toast.

Change-Id: Ia74c7d0622446464a2a57395cf3f336eeb172ca3
Signed-off-by: iceandfire <arhamjamal@gmail.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/QuickSettings.java