hwui: fix projection uniform caching.
The previous caching would only consider a change in the
actual matrix to figure out that the projection matrix
uniform needs to be updated. This would not work if the
requested projection matrix is identical, but only the
offsetting would change.

The fix now considers changes in the requested offsetting.

Change-Id: I49848de99a4ef80f43d9ca8437aa8403759c7d28