Only one task within the stack should moveToBack
In case of the last stack of application move to back,
ONLY one stack within stack should move to back.
but stack itseft move to back.

Default bahavior of ActivityManager is that the stack will
go back within stack when the last stack of application move
to back. The stack is managed by DisplayContent of WindowManagerService
should be same but taking one more action that move stack itself.

Change-Id: If394ddf16b0fcda2ae3252e8f3b5336322ef6808
Signed-off-by: Seunghyun Lee <seunghyun85.lee@lge.com>