systemui: adjust position of clear recents button
If there is a statusbar then align to it. Otherwise aling to task views rather than top/right.
This patch fixed the position in tablets (align it with the search bar) and in landscape mode
(reducing the space between task views and the clear recents button), with is a more consistent
ui because the button is always properly aligned with the recent task views.

Change-Id: Ic867d5421cacba627c4b844db3d3715f20a4444c
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>