Heads up snooze tweaks
Heads up has a snooze function in 5.1.1 but a few know about it. If you swipe up from heads up it will dismiss the notification in dimissed package and heads up will not show for 1 minute. The same happens if you try to expand the status bar from a heads up notification (not a smart behavior).
Now a toast message will be showed to inform the user and heads up will only be dismissed in swipe up action

Change-Id: I21d5920888393c11079b7602ab87fa728a659b61

Conflicts:
	packages/SystemUI/res/values/cr_strings.xml