Lockscreen: Fix incorrent behavior on camera shortcut
Edge case:
This bug is only visible after a clean flash, without gapps (or, at least, without a Camera installed). The camera shortcut is hidden correctly (default behavior), but if you try to set it to "none" in the Settings --> Security --> Lockscreen Shortcuts preference screen, then the camera shortcut is visible. This fixes that, without any other change in behavior.